==> 02-literature.Rmd <==
# Fuentes


Las fuentes de información serán anonimizadas en Python por me dio de la biblioteca “UUID” el cual lo relaciona con un identificador único, y luego se puede relación con el nombre original para uso interno de la operadora 

Las fuentes de información son las siguientes:

**Nokia:**  Se tiene  la información del tráfico ( Payload)  por cada estación base con periodicidad diaria desde Enero 2021 a febrero 2023

**Dane:** Información de municipios de la cantidad de población , los municipios seleccionados fueron los categorizados por el dane como de "poblacion media"

**Paginas Web:** Información de las fechas de las ferias y fiestas de los municipios de Cundinamarca
